requirements.txt
PyYAML==3.10
docker-py==1.2.3
dockerpty==0.3.4
docopt==0.6.1
pytimeparse==1.1.2
requests==2.6.1
retrying==1.2.2
six==1.7.3
texttable==0.8.2
websocket-client==0.11.0
